+++ workspaces/bulk-import/plugins/bulk-import-common/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## @red-hat-developer-hub/backstage-plugin-bulk-import-common [0.2.0](https://github.com/janus-idp/backstage-plugins/compare/@red-hat-developer-hub/backstage-plugin-bulk-import-common@0.1.0...@red-hat-developer-hub/backstage-plugin-bulk-import-common@0.2.0) (2024-07-25)
 
+## 6.8.0
+
+### Minor Changes
+
+- 0739b9a: Backstage version bump to v1.45.2
+
 ## 6.7.2
 
 ## 6.7.1