+++ workspaces/lightspeed/plugins/lightspeed-backend/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @red-hat-developer-hub/backstage-plugin-lightspeed-backend
 
+## 1.1.1
+
+### Patch Changes
+
+- b0c55d6: Updated dependency `msw` to `2.12.4`.
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-common@1.1.1
+
+## 1.1.0
+
+### Minor Changes
+
+- 99f35d5: Backstage version bump to v1.45.2
+
+### Patch Changes
+
+- Updated dependencies [99f35d5]
+  - @red-hat-developer-hub/backstage-plugin-lightspeed-common@1.1.0
+
 ## 1.0.4
 
 ### Patch Changes